requirements.txt
discord-py==2.6.3
discord-typings==0.9.0
python-dotenv==1.1.0
